+++ src/ejercicio06/VolumenCilindro.java
@@ -2,7 +2,7 @@ package ejercicio06;
 
 public class VolumenCilindro {
 
-	public static void volumencilindro(int altura, int radio, int opcion) {
+	public static void volumencilindro(double altura, double radio, int opcion) {
 
 		System.out.println("Opcion elegida: 2. Volumen del Cilindro");
 		System.out.println(Math.PI * Math.pow(radio, 2) * altura);
